Move visualize into own module
